# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenRelay API",
    "description": "Form orchestration backend.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/oauth/callback": {
      "get": {
        "tags": [
          "oauth"
        ],
        "operationId": "callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error_description",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Redirect to admin SPA with token (sign-in) or status (link)"
          },
          "400": {
            "description": "Missing/invalid state or code"
          },
          "404": {
            "description": "OAuth not configured"
          },
          "502": {
            "description": "Token exchange failed"
          }
        }
      }
    },
...
  },
...
}